добавил тесты

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -14,17 +14,16 @@ public class Film {
     private int id;
     @NotBlank(message = "Название не может быть пустым")
     private String name;
-
     @Size(max = 200, message = "Максимальная длина описания — 200 символов")
     private String description;
     String releaseDate;
     @Positive(message = "Продолжительность фильма должна быть положительной")
     private Long duration;
 
-    public Film(String name, String description, String releaseDate, Long duration) {
+    public Film(String name, String description, String releaseDate, int duration) {
         this.name = name;
         this.description = description;
         this.releaseDate = releaseDate;
-        this.duration = duration;
+        this.duration = Long.valueOf(duration);
     }
 }
